+++ example/index.tsx
@@ -11,10 +11,12 @@ import {
     createSlide,
     createCounter,
     createBoard,
+    createIframe,
 } from "./apps";
 import "../dist/style.css";
 import "./register";
 import "./index.css";
+import { DefaultHotKeys } from "white-web-sdk";
 
 const sdk = new WhiteWebSdk({
     appIdentifier: import.meta.env.VITE_APPID,
@@ -150,6 +152,13 @@ const joinRoom = ref => {
                 disableMagixEventDispatchLimit: true,
                 disableNewPencil: false,
                 floatBar: true,
+                hotKeys: {
+                    ...DefaultHotKeys,
+                    changeToClick: "c",
+                    changeToSelector: "s",
+                    changeToPencil: "p",
+                    changeToEraser: "e",
+                },
             })
             .then(async room => {
                 (window as any).room = room;
@@ -188,6 +197,7 @@ const App = () => {
         joinRoom(ref.current).then(() => {
             if (manager) {
                 setPageState(manager.pageState);
+                createIframe(manager);
                 return manager.emitter.on("pageStateChange", state => {
                     setPageState(state);
                 });
@@ -227,6 +237,9 @@ const App = () => {
                 </button>
                 <button className="side-button" onClick={() => createVideo(manager)}>
                     视频
+                </button>
+                <button className="side-button" onClick={() => createIframe(manager)}>
+                    Iframe Bridge
                 </button>
                 <button className="side-button" onClick={replay}>
                     回放